Port core tests for XPU.
