+++ .github/workflows/build_and_test.yml
@@ -58,7 +58,7 @@ jobs:
         if: ${{ env.BACKEND == 'XPU'}}
         run: |
           cd python/test/unit/language
-          python3 -m pytest --verbose --device xpu test_core_xpu.py
+          python3 -m pytest --verbose --device xpu test_core.py
 
       - name: Run assert/print tests
         if: ${{ env.BACKEND == 'XPU'}}
