+++ scripts/test-triton.sh
@@ -100,7 +100,7 @@ function run_core_tests {
     echo "Not found '${CORE_TEST_DIR}'. Build Triton please" ; exit 3
   fi
   cd $CORE_TEST_DIR
-  python3 -m pytest --verbose --device xpu test_core_xpu.py
+  python3 -m pytest --verbose --device xpu test_core.py
   if [ $? -ne 0 ]; then
     echo "FAILED: return code $?" ; exit $?
   fi
